requirements.txt
python-telegram-bot==20.6
openai==1.3.5
python-dotenv==1.0.1
google-generativeai==0.3.1
notion-client==2.0.0
aiohttp==3.9.1
aiofiles==23.2.1
openai
telegram